requirements.txt
fastmcp
sqlite3-binary
